requirements.txt
streamlit
requests
openai
streamlit-audiorecorder
soundfile
